% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absoluteTest.genePvals.R
\name{absoluteTest.genePvals}
\alias{absoluteTest.genePvals}
\title{A function to calculate the P values for each gene as compared with either  0, the gene set mean, or the full PDF of the gene set.}
\usage{
absoluteTest.genePvals(QSarray, path.index = 1:numPathways(QSarray),
  silent = TRUE, FastApproximated = TRUE, addVIF = !is.null(QSarray$vif),
  NPoints = QSarray$n.points/8, compareTo = c("zero", "mean", "pdf"))
}
\arguments{
\item{QSarray}{A QSarray object, as generated by aggregateGeneSet and possibly modified by calcVIF}

\item{path.index}{The pathways to calculate the pVals for.}

\item{silent}{If false, print a "." every fifth pathway, as a way to keep track of progress}

\item{FastApproximated}{fast mode which uses normal approximations for the PDF's}

\item{addVIF}{Whether to include the VIF in the calculation of the p-values. By default, if the VIF has been calculated, it will be used}

\item{NPoints}{points to use}

\item{compareTo}{the null hypothesis that each gene set is tested against.}
}
\value{
gene set associated P values
}
\description{
A function to calculate the P values for each gene as compared with either  0, the gene set mean, or the full PDF of the gene set.
}

