% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absoluteTest.genePvalsFAST.R
\name{absoluteTest.genePvalsFAST}
\alias{absoluteTest.genePvalsFAST}
\title{function to calculate the homogeneity P value faster}
\usage{
absoluteTest.genePvalsFAST(QSarray, CompareWithZero = TRUE)
}
\arguments{
\item{QSarray}{A QSarray object, as generated by aggregateGeneSet and possibly modified by calcVIF}

\item{CompareWithZero}{Logical, if TRUE compares with mean of zero, else with mean of pathway}
}
\value{
p values faster
}
\description{
function to calculate the homogeneity P value faster
}

