% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absoluteTest.R
\name{absoluteTest}
\alias{absoluteTest}
\title{absolute test homogeneity score}
\usage{
absoluteTest(eset, QSarray, p.adjust.method = "fdr", silent = F)
}
\arguments{
\item{eset}{expression set matrix}

\item{QSarray}{qusage object}

\item{p.adjust.method}{method for correcting falses}

\item{silent}{verbose}
}
\value{
homogeneity score
}
\description{
absolute test homogeneity score
}

