% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGeneSetArm.R
\name{aggregateGeneSetArm}
\alias{aggregateGeneSetArm}
\title{Combine individual gene differential expresseion for each pathway (Neg) ~ 1 minute}
\usage{
aggregateGeneSetArm(geneResults, geneSets, n.points = 2^12, silent = TRUE)
}
\arguments{
\item{geneResults}{A QSarray object, as generated by makeComparison}

\item{geneSets}{a list of pathways to be compared, each item in the list is a vector of names that correspond to the gene names from Baseline/PostTreatment}

\item{n.points}{the number of points to sample the convoluted t-distribution at.}

\item{silent}{If false, print a "." every fifth pathway, as a way to keep track of progress}
}
\value{
gene set aggregations
}
\description{
Combine individual gene differential expresseion for each pathway (Neg) ~ 1 minute
}

