% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximatedNu.R
\docType{data}
\name{approximatedNu}
\alias{approximatedNu}
\title{Nu values approximated, this is a vector holding constants}
\format{An object of class \code{numeric} of length 271.}
\usage{
approximatedNu
}
\value{
vector of values for Nu
}
\description{
Nu values approximated, this is a vector holding constants
}
\keyword{datasets}

