% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calIndividualExpressions.R
\name{calcIndividualExpressions}
\alias{calcIndividualExpressions}
\title{script from Qusage, used for testing}
\usage{
calcIndividualExpressions(Baseline, PostTreatment, paired = FALSE,
  min.variance.factor = 10^-6, na.rm = TRUE)
}
\arguments{
\item{Baseline}{is the matix of gene expressions at baseline, row names are gene names}

\item{PostTreatment}{is the matix of gene expressions after treatment, row names are gene names}

\item{min.variance.factor}{error numeric}

\item{na.rm}{boolean for removing na terms}

\item{paired:}{logical, whether the data is paired or not}
}
\value{
qusage array object
}
\description{
script from Qusage, used for testing
}

