% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndividualExpressionsArm.R
\name{calcIndividualExpressionsArm}
\alias{calcIndividualExpressionsArm}
\title{modificed expression calculation that enforces clean data without NAs for speeding up qusage}
\usage{
calcIndividualExpressionsArm(Baseline, PostTreatment, paired = FALSE,
  min.variance.factor = 10^-6)
}
\arguments{
\item{Baseline}{expression values before treatment}

\item{PostTreatment}{expression values after treatment}

\item{paired}{boolean, paired reads}

\item{min.variance.factor}{numeric, variance factor for pooled variance}
}
\description{
modificed expression calculation that enforces clean data without NAs for speeding up qusage
}

