% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndividualExpressionsC.R
\name{calcIndividualExpressionsC}
\alias{calcIndividualExpressionsC}
\title{modificed expression calculation that enforces clean data without NAs for speeding up qusage}
\usage{
calcIndividualExpressionsC(Baseline, PostTreatment, paired = FALSE,
  min.variance.factor = 10^-6, no.cores = 3)
}
\arguments{
\item{Baseline}{expression values before treatment}

\item{PostTreatment}{expression values after treatment}

\item{paired}{boolean, paired reads}

\item{min.variance.factor}{numeric, variance factor for pooled variance}

\item{no.cores}{integer, number of cores used in parallel}
}
\description{
modificed expression calculation that enforces clean data without NAs for speeding up qusage
}

