% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPCor.R
\name{calcPCor}
\alias{calcPCor}
\title{A function to calculate the Variance Inflation Factor (VIF) for each of the gene sets input in geneSets This function depends on which method you used to calculate the variance originally.  If you assumed pooled variance, then the variance will be calculated using LIMMA's "interGeneCorrelation" method (see Wu and Smyth, Nucleic Acids Res. 2012). Otherwise, the method will calculate the VIF separately for each group and return the average of each group's vif.}
\usage{
calcPCor(eset, geneResults, useCAMERA = geneResults$var.method == "Pooled",
  useAllData = TRUE)
}
\arguments{
\item{eset}{a matrix of log2(expression values). This must be the same dataset that was used to create geneResults}

\item{geneResults}{A QSarray object, as generated by either makeComparison or aggregateGeneSet}

\item{useCAMERA}{The method used to calculate variance. See the description for more details. By default, it uses the parameter stored in the geneResults}

\item{useAllData}{Boolean parameter determining whether to use all data in eset to calculated the VIF, or whether to only use data from the groups being contrasted. Only used if useCAMERA==FALSE}
}
\value{
correlation of vif between gene sets
}
\description{
A function to calculate the Variance Inflation Factor (VIF) for each of the gene sets input in geneSets This function depends on which method you used to calculate the variance originally.  If you assumed pooled variance, then the variance will be calculated using LIMMA's "interGeneCorrelation" method (see Wu and Smyth, Nucleic Acids Res. 2012). Otherwise, the method will calculate the VIF separately for each group and return the average of each group's vif.
}

