% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTwoDistsFaster.R
\name{compareTwoDistsFaster}
\alias{compareTwoDistsFaster}
\title{Compute p-value of two distributions not centered at zero}
\usage{
compareTwoDistsFaster(dens1 = runif(256 * 8, 0, 1), dens2 = runif(256 * 8,
  0, 1), alternative = "two.sided")
}
\arguments{
\item{dens1}{density 1}

\item{dens2}{density 2}

\item{alternative}{alternative pdf}
}
\value{
compared pdf
}
\description{
Compute p-value of two distributions not centered at zero
}

