% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_table.R
\name{course_table}
\alias{course_table}
\title{Creates html table of course contents with trailing opacity set by week n}
\usage{
course_table(b1, b2, b1list, b2list, week)
}
\arguments{
\item{b1}{character string for name of block 1}

\item{b2}{character string for name of block 2}

\item{b1list}{vector (5L) of character strings of lecture titles for block 1}

\item{b2list}{vector (5L) of character strings of lecture titles for block 2}

\item{week}{week (1 to 10) of semester up to which opacity = 1}
}
\description{
Creates html table of course contents with trailing opacity set by week n
}
\examples{
block1_name = "multilevel modelling<br>working with group structured data"
block1_lecs = c("regression refresher","introducing multilevel models","more multilevel models","more complex groupings","recap")
block2_name = "factor analysis<br>working with multi-item measures"
block2_lecs = c("PCA","EFA","EFA 2","CFA","CFA 2")

# put this line in a slide chunk and use results = 'asis'
# make two column layout for full year courses
course_table(block1_name, block2_name, block1_lecs, block2_lecs, week=4)
}
