% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGenes.R
\name{filterGenes}
\alias{filterGenes}
\title{Filter out genes with SD lower than a threshold and mean lower than another threshold}
\usage{
filterGenes(QSarray, Min_SD = 0.01, Min_Mean = 0)
}
\arguments{
\item{QSarray}{A QSarray object, as generated by makeComparison}

\item{Min_SD}{threshold for SD}

\item{Min_Mean}{threshold for mean}
}
\value{
qsuage array object filtered
}
\description{
Filter out genes with SD lower than a threshold and mean lower than another threshold
}

