% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneityScore.R
\name{homogeneityScore}
\alias{homogeneityScore}
\title{homogenity score}
\usage{
homogeneityScore(QSarray)
}
\arguments{
\item{Qsarray}{qusage object container}
}
\value{
score of homogeneity
}
\description{
homogenity score
}

