% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertcallout.R
\name{insertcallout}
\alias{insertcallout}
\title{Insert callout}
\usage{
insertcallout()
}
\description{
Call this function as an addin to insert callout code at the cursor position.
Can add as a keyboard shortcut (I use ctrl + shift + q)
}
