% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertcolumns.R
\name{insertcolumns}
\alias{insertcolumns}
\title{Insert columns}
\usage{
insertcolumns()
}
\description{
Call this function as an addin to insert 2 column code at the cursor position.
Can add as a keyboard shortcut (I use ctrl + shift + q or e)
}
