% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertlinkout.R
\name{insertlinkout}
\alias{insertlinkout}
\title{Insert linkout}
\usage{
insertlinkout()
}
\description{
Call this function as an addin to insert blank link at the cursor position.
Can add as a keyboard shortcut
}
