% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeComparisonArm.R
\name{makeComparisonArm}
\alias{makeComparisonArm}
\title{A wrapper function to calculate comparisons between groups in a dataset. A note on var.equal: LIMMA's linear model function can only be run when assuming equal  variances. If var.equal==TRUE, then a linear model will be created on the entire dataset at once. Otherwise, calcIndividualExpressions will be called for the comparison of interest, and means and SDs will be calculated directly.One benefit of using LIMMA's pooled variance calculation is that the linear models allow for more complicated comparisons (e.g. "(A+B)-C" or similar). This may be of interest to some users, but in order to do this, you must assume equal variances between all groups. One caveat regarding paired samples: LIMMA can not fit a linear model when the paired samples are convoluted with the groups (e.g. one set of paired (trt vs mock) samples in patients with  disease, combined with a set of paired samples from healthy controls). If var.equal==TRUE, these groups must be run separately to correctly fit the model (e.g. run disease first, then healthy controls).}
\usage{
makeComparisonArm(eset, labels, contrast, pairVector = NULL,
  var.equal = FALSE, bayesEstimation = TRUE, min.variance.factor = 10^-8)
}
\arguments{
\item{eset}{a matrix of log2(expression values), with rows of features and columns of samples}

\item{labels}{vector of labels representing each column of eset.}

\item{contrast}{a string describing which of the groups in 'labels' we want to compare. This is usually of the form 'trt-ctrl', where 'trt' and 'ctrl' are groups represented in 'labels'.}

\item{pairVector}{A vector of factors (usually just 1,2,3,etc.) describing the sample pairings. This is often just a vector of patient IDs or something similar. If not provided, all samples are assumed to be independent.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch approximation is used.}

\item{bayesEstimation}{if true, use a bayesian framework to estimate the standard deviation (via limma's eBayes function).}

\item{min.variance.factor}{a factor to add to the SDs to ensure that none are equal to 0. Only used if var.equal==FALSE or bayesEstimation==FALSE.}

\item{implicitContrast}{boolean to explicitly contrast the design fit using an intercept constant, this increases for performance by minimizing calls to limma}
}
\value{
qusage data object
}
\description{
A wrapper function to calculate comparisons between groups in a dataset. A note on var.equal: LIMMA's linear model function can only be run when assuming equal  variances. If var.equal==TRUE, then a linear model will be created on the entire dataset at once. Otherwise, calcIndividualExpressions will be called for the comparison of interest, and means and SDs will be calculated directly.One benefit of using LIMMA's pooled variance calculation is that the linear models allow for more complicated comparisons (e.g. "(A+B)-C" or similar). This may be of interest to some users, but in order to do this, you must assume equal variances between all groups. One caveat regarding paired samples: LIMMA can not fit a linear model when the paired samples are convoluted with the groups (e.g. one set of paired (trt vs mock) samples in patients with  disease, combined with a set of paired samples from healthy controls). If var.equal==TRUE, these groups must be run separately to correctly fit the model (e.g. run disease first, then healthy controls).
}

