% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeComparisonC.R
\name{makeComparisonC}
\alias{makeComparisonC}
\title{more efficient and faster comparison function}
\usage{
makeComparisonC(eset, labels, contrast, pairVector = NULL,
  var.equal = FALSE, bayesEstimation = TRUE, min.variance.factor = 10^-8)
}
\arguments{
\item{eset}{a matrix of log2(expression values), with rows of features and columns of samples}

\item{labels}{vector of labels representing each column of eset}

\item{contrast}{a string describing which of the groups in 'labels' we want to compare. This is usually of the form 'trt-ctrl', where 'trt' and 'ctrl' are groups represented in 'labels'}

\item{pairVector}{A vector of factors (usually just 1,2,3,etc.) describing the sample pairings. This is often just a vector of patient IDs or something similar. If not provided, all samples are assumed to be independent.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch approximation is used.}

\item{bayesEstimation}{if true, use a bayesian framework to estimate the standard deviation (via limma's eBayes function)}
}
\value{
a Qusage array object
}
\description{
more efficient and faster comparison function
}

