% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeqms.R
\name{makeqms}
\alias{makeqms}
\title{make quick marks for turnitin.
csv should have two columns, the first is the name of the quickmark
the second is the content.
should be able to use html in the content if you need it for bold/ital etc.}
\usage{
makeqms(csvfile, setname = "myquickmarkset")
}
\description{
make quick marks for turnitin.
csv should have two columns, the first is the name of the quickmark
the second is the content.
should be able to use html in the content if you need it for bold/ital etc.
}
