% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ni.R
\name{Ni}
\alias{Ni}
\title{helper funciton to calculate DOF}
\usage{
Ni(SigmaOld, SigmaYoung, NOld, NYoung)
}
\arguments{
\item{SigmaOld}{first sigma to compare}

\item{SigmaYoung}{second sigma to compare}

\item{NOld}{Non-nas of old, ncolumns}

\item{NYoung}{non-nas of columns of young}
}
\value{
degrees of freedom
}
\description{
helper funciton to calculate DOF
}

