% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfScaleFactor.R
\name{pdfScaleFactor}
\alias{pdfScaleFactor}
\title{calculate a scaling factor to multiply the PDF by (so that it actually integrates to 1)}
\usage{
pdfScaleFactor(QSarray, addVIF = !is.null(QSarray$vif))
}
\arguments{
\item{QSarray}{qusage object}

\item{addVIF}{vif value}
}
\value{
pdf plot
}
\description{
calculate a scaling factor to multiply the PDF by (so that it actually integrates to 1)
}

