% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscale.R
\name{pscale}
\alias{pscale}
\title{round to 20 point psychology marking scale.
will round to nearest point on scale. if mid-point, will round up.}
\usage{
pscale(x, lgrade = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{lgrade}{logical specifying whether letter grades are desired (default = FALSE)}
}
\description{
round to 20 point psychology marking scale.
will round to nearest point on scale. if mid-point, will round up.
}
\examples{
pscale(c(46.5,34,76.6))
[1] 48 32 78
}
