% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qusageCutArm.R
\name{qusageCutArm}
\alias{qusageCutArm}
\title{This file contains the primary code for the qusage methodology Author: Christopher Bolen Gur Yaari (c) 2013 Yale University. All rights reserved. but runs welch's approximation in C++.}
\usage{
qusageCutArm(eset, labels, contrast, geneSets, pairVector = NULL,
  var.equal = FALSE, filter.genes = FALSE, n.points = 2^12)
}
\arguments{
\item{eset}{a matrix of log2(expression values), with rows of features and columns of samples. OR an object of class ExpressionSet}

\item{labels}{vector of labels representing each column of eset.}

\item{contrast}{a string describing which of the groups in 'labels' we want to compare. This is usually of the form 'trt-ctrl', where 'trt' and 'ctrl' are groups represented in 'labels'.}

\item{geneSets}{a list of pathways to be compared. Each item in the list is a vector of names that correspond to the row names of eset.}

\item{pairVector}{A vector of factors (usually just 1,2,3,etc.) describing the sample pairings. This is often just a vector of patient IDs or something similar. If not provided, all samples are assumed to be independent.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch approximation is used.}

\item{filter.genes}{a boolean indicating whether the genes in eset should be filtered to remove genes with low mean and sd.}

\item{n.points}{The number of points to sample the convolution at. Passed to aggregateGeneSet}
}
\value{
qusage data array for welch's approx
}
\description{
This file contains the primary code for the qusage methodology Author: Christopher Bolen Gur Yaari (c) 2013 Yale University. All rights reserved. but runs welch's approximation in C++.
}

