% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_basicrs.R
\name{sim_basicrs}
\alias{sim_basicrs}
\title{Quickly sim data for a basic model y~1+x+(1+x|g)}
\usage{
sim_basicrs(seed = NULL, b0 = 0, b1 = 1, z0 = 1, z1 = 1, e = 1)
}
\arguments{
\item{seed}{random seed to use}

\item{b0}{fixed intercept}

\item{b1}{fixed slope}

\item{z0}{random intercept sd}

\item{z1}{random slope sd}

\item{e}{residual var}
}
\description{
Quickly sim data for a basic model y~1+x+(1+x|g)
}
\examples{
sim_basicrs()
library(lme4)
summary(lmer(y~1+x+(1+x|g),data=sim_basicrs()))
}
