% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoepsy_turnitin.R
\name{uoepsy_turnitin}
\alias{uoepsy_turnitin}
\title{Create a spreadsheet with turnitin links from a set of files.}
\usage{
uoepsy_turnitin(submission_dir, output_file, rename_files = FALSE)
}
\arguments{
\item{submission_dir}{directory holding all and only the submitted files}

\item{output_file}{filepath to output csv}

\item{rename_files}{if TRUE will rename files to just the exam numbers (where possible)}
}
\description{
Create a spreadsheet with turnitin links from a set of files.
}
\examples{
# from turnitin submission box, download all files, and unzip to a folder, e.g. dapr1reps/
uoepsy_turnitin(submission_dir = "~/Desktop/dapr1reps/", output_file = "~/Desktop/dapr_marks.csv")
}
