% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_nonote_render.R
\name{xaringan_nonote_render}
\alias{xaringan_nonote_render}
\title{render xaringan slides without presenter notes}
\usage{
xaringan_nonote_render(fpath = NULL, outdir = NULL)
}
\arguments{
\item{fpath}{path to .Rmd file}

\item{outdir}{directory to place rendered html}
}
\description{
render xaringan slides without presenter notes
}
