% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoepsy_turnitin.R
\name{uoepsy_turnitin}
\alias{uoepsy_turnitin}
\title{make spreadsheet with turnitin links from set of files}
\usage{
uoepsy_turnitin(submission_dir, output_file, seed = FALSE, rename_files = TRUE)
}
\arguments{
\item{submission_dir}{directory holding all and only the submitted files}

\item{output_file}{filepath to output csv}

\item{seed}{search script for seed? (looks for x<-, based on dapr labtests)}

\item{rename_files}{if TRUE will rename files to just exam numbers (where possible)}
}
\description{
make spreadsheet with turnitin links from set of files
}
\examples{
uoepsy_turnitin("~/Desktop/dapr1reps/","~/Desktop/dapr_marks.csv")
}
